[feat] 테이블 분리 #143
