+++ src/components/dialogs/TermsDialog.tsx
