+++ src/pages/WorkspacePage.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 
-import MainHeader from '@/components/header/MainHeader';
-import { HistoryTable } from '@/components/tables/HistoryTable';
-import HomePopup from '@/components/workspace/HomePopup';
-import RecentProject from '@/components/workspace/RecontProject';
+import HomePopup from '@/components/guide/HomePopup';
+import RecentProject from '@/components/section/contents/RecontProject';
+import MainHeader from '@/components/section/header/MainHeader';
+import { HistoryTable } from '@/components/tables/workspace/HistoryTable';
 import jisuImage from '@/images/avatar/jisu.jpg';
 
 const WorkspacePage = () => {
